requirements.txt
fastapi
uvicorn
pytesseract
pdf2image
pillow
spacy
python-multipart
pdfplumber